requirements.txt
boto3
botocore
Pillow
img2pdf
fastapi
uvicorn
pydantic-settings
pytest
python-multipart 